refactor sdk internal routes into route modules
